requirements.txt
asgiref==3.11.0
Django==6.0.1
django-extensions==4.1
djangorestframework==3.16.1
gunicorn==24.1.1
packaging==26.0
sqlparse==0.5.5
whitenoise==6.11.0
openai
python-dotenv
google-genai
django-recaptcha
